Refactor custom ops into proper file locations
