+++ QEfficient/customop/rms_norm.py
@@ -1,61 +1,51 @@
 # -----------------------------------------------------------------------------
 #
-# Copyright (c)  2023-2024 Qualcomm Innovation Center, Inc. All rights reserved.
+# Copyright (c) 2023-2024 Qualcomm Innovation Center, Inc. All rights reserved.
 # SPDX-License-Identifier: BSD-3-Clause
 #
 # -----------------------------------------------------------------------------
 
-"""
-RMS Norm CustomOp Node in QAic Domain for Cloud AI 100
-This is to handle the FP16 Overflow seen in RMS Norm for LLMs
-"""
-
+import onnxscript
 import torch
-from torch.onnx.symbolic_helper import parse_args
+from torch import nn
 
-op_source = """
-#include <torch/script.h>
-
-torch::Tensor custom_rms_norm(torch::Tensor hidden_states, torch::Tensor weight, double eps) {
-  torch::Tensor output;
-  torch::Tensor variance;
-  bool keepdim;
-  // double eps = 1e-5;
-  variance = hidden_states.pow(2).mean(-1, keepdim=true);
-  output = hidden_states * torch::rsqrt(variance + eps);
-  output = output * weight;
-  return output;
-}
-
-TORCH_LIBRARY(QAic, m) {
-  m.def("QEffCustomRMSNorm", &custom_rms_norm);
-}
-"""
-
-# Compile and load the custom op
-torch.utils.cpp_extension.load_inline(
-    name="custom_rms_norm",
-    cpp_sources=op_source,
-    is_python_module=False,
-    verbose=True,
-)
+ops = onnxscript.opset13
 
 
-# Wrapper module for custom relu C++ op
-class QEffCustomRMSNorm(torch.nn.Module):
-    def __init__(self, normalized_shape, eps=1e-5, device=None, dtype=None, **kwargs):
-        super().__init__()
-        self.weight = torch.nn.Parameter(torch.empty(normalized_shape, device=device, dtype=dtype))
-        self.eps = eps
+@onnxscript.script(onnxscript.values.Opset(domain="com.qti.aisw.onnx", version=1))
+def CustomRMSNorm(hidden_states: onnxscript.FLOAT, weight: onnxscript.FLOAT, epsilon: float):
+    weight = ops.Cast(weight, to=1)
+    variance = ops.ReduceMean(ops.Pow(hidden_states, 2), axes=[-1], keepdims=1)
+    epsilon = ops.Expand(epsilon, ops.Shape(variance))
+    hidden_states = hidden_states * ops.Reciprocal(ops.Sqrt(variance + epsilon))
+    return weight * hidden_states
+
+
+class CustomRMSNormFunc(torch.autograd.Function):
+    @staticmethod
+    def forward(hidden_states: torch.Tensor, weight: torch.Tensor, epsilon: float):
+        variance = hidden_states.pow(2).mean(-1, keepdim=True)
+        hidden_states = hidden_states * torch.rsqrt(variance + epsilon)
+        return weight * hidden_states
+
+    @staticmethod
+    def setup_context(ctx, inputs, outputs):
+        pass
+
+    @staticmethod
+    def symbolic(g: torch.Graph, hidden_states: torch.Value, weight: torch.Value, epsilon: torch.Value) -> torch.Value:
+        return g.onnxscript_op(CustomRMSNorm, hidden_states, weight, epsilon_f=epsilon).setTypeAs(hidden_states)
+
+
+class CustomRMSNormAIC(nn.Module):
+    """
+    RMSNorm module that works by replacing the current module with compiler known custom-op.
+    """
+
+    def __init__(self, hidden_size, eps=1e-05):
+        super(CustomRMSNormAIC, self).__init__()
+        self.variance_epsilon = eps
+        self.weight = torch.nn.Parameter(torch.ones(hidden_size))
 
     def forward(self, hidden_states):
-        return torch.ops.QAic.QEffCustomRMSNorm(hidden_states, self.weight, self.eps)
-
-
-# ONNX export symbolic helper
-@parse_args("v", "v", "f")
-def custom_rms_norm(g, hidden_states, weight, eps):
-    return g.op("QAic::QEffCustomRMSNorm", hidden_states, weight, eps_f=eps).setTypeAs(hidden_states)
-
-
-torch.onnx.register_custom_op_symbolic("QAic::QEffCustomRMSNorm", custom_rms_norm, 1)
+        return CustomRMSNormFunc.apply(hidden_states, self.weight, self.variance_epsilon)
