+++ QEfficient/customop/rms_norm_native.yaml
@@ -1,6 +1,6 @@
 # -----------------------------------------------------------------------------
 #
-#Copyright (c)  2023-2024 Qualcomm Innovation Center, Inc. All rights reserved.
+#Copyright (c) 2023-2024 Qualcomm Innovation Center, Inc. All rights reserved.
 #SPDX-License-Identifier: BSD-3-Clause
 #
 # -----------------------------------------------------------------------------
